+++ lib/togls/errors.rb
@@ -1,6 +1,7 @@
 module Togls
-  class NoFeaturesError < StandardError; end
-  class MissingDriver < StandardError; end
-  class InvalidDriver < StandardError; end
-  class NotImplemented < StandardError; end
+  class Error < StandardError; end
+  class NoFeaturesError < Error; end
+  class MissingDriver < Error; end
+  class InvalidDriver < Error; end
+  class NotImplemented < Error; end
 end